Moved ForwardedAttachmentsDownloaderWorker to jetpack/workmanager.| #891
