+++ FlowCrypt/src/main/java/com/flowcrypt/email/jobscheduler/JobIdManager.kt
@@ -27,8 +27,6 @@ class JobIdManager {
     const val JOB_TYPE_EMAIL_AND_NAME_UPDATE = 3
     const val JOB_TYPE_PREPARE_OUT_GOING_MESSAGE = 4
     const val JOB_TYPE_SEND_MESSAGES = 5
-    const val JOB_TYPE_DOWNLOAD_FORWARDED_ATTACHMENTS = 6
     const val JOB_TYPE_FEEDBACK = 6
-    const val JOB_TYPE_MOVE_MESSAGES = 7
   }
 }
